とりあえずgraphiqlが動いたところまで

+++ backend/migrations/2020-05-11-000300_create_clients/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `clients` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `id` int(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(255) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4
